pyproject.toml
[project]
name = "think-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "pgvector>=0.4.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "supabase>=2.28.2",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.6",
]

[tool.ruff]
exclude = ["supabase/database_types.py"]
line-length = 100
target-version = "py312"

[tool.ruff.format]
exclude = ["supabase/database_types.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "PT"]

[tool.mypy]
python_version = "3.12"
strict = true
exclude = ["supabase/database_types.py"]
